+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
+
+- Int Array declaration -> `x: Int[4] = [1, 2, 3, 4]`
+- Array Access -> `y: Int = x[0]`
+
 ### Changed
 ### Fixed